Use new emptystate
