--- src/components/empty/empty.1.example
@@ -1,14 +1,0 @@
-html!{
-    <EmptyState
-        title="Empty state"
-        icon={Icon::Cubes}
-        primary={Action::new("Push me", ctx.link().callback(|_|{}))}
-        secondaries={vec![
-            Action::new("Try me", ctx.link().callback(|_|{})),
-            Action::new("Me too", ctx.link().callback(|_|{})),
-            Action::new("Here, here", ctx.link().callback(|_|{})),
-        ]}
-        >
-        {"This section should explain why the state is empty, and what you can do next."}
-    </EmptyState>
-}
