+++ src/components/empty_state/empty_state.6.example
@@ -1,0 +1,11 @@
+html!{
+    <EmptyState>
+        <EmptyStateHeader
+            title_text={html!("Loading")}
+            heading_level={EmptyStateHeadingLevel::H4}
+            icon={html!(
+                <EmptyStateIcon icon={html!(<Spinner/>)} />
+            )}
+        />
+    </EmptyState>
+}
