+++ src/components/mod.rs
@@ -14,7 +14,7 @@ mod context_selector;
 mod description_list;
 mod divider;
 mod dropdown;
-mod empty;
+mod empty_state;
 mod expandable_section;
 mod file_upload;
 mod form;
@@ -56,7 +56,7 @@ pub use context_selector::*;
 pub use description_list::*;
 pub use divider::*;
 pub use dropdown::*;
-pub use empty::*;
+pub use empty_state::*;
 pub use expandable_section::*;
 pub use file_upload::*;
 pub use form::*;
